test: 지원 성적 제출 테스트 추가

+++ src/main/java/com/example/solidconnection/application/domain/Gpa.java
@@ -13,10 +13,10 @@ import lombok.NoArgsConstructor;
 public class Gpa {
 
     @Column(nullable = false, name = "gpa")
-    private Float gpa;
+    private Double gpa;
 
     @Column(nullable = false, name = "gpa_creteria")
-    private Float gpaCriteria;
+    private Double gpaCriteria;
 
     @Column(nullable = false, name = "gpa_report_url", length = 500)
     private String gpaReportUrl;
